Optimizes `log2` and enables BMI/LZCNT extensions (#159)
* Enable BMI1 instructions for AVX2 and AVX512 variants

Adds BMI1 support to optimized x86_64 builds by explicitly including the necessary flags for both MSVC and GCC/Clang. This ensures that the full range of bit manipulation instructions is available alongside BMI2 when targeting AVX2 or AVX512 architectures.

* Optimize zxc_log2_u32 with bit manipulation intrinsics

Replaces the manual loop implementation of `zxc_log2_u32` with platform-specific intrinsics (`_BitScanReverse` and `__builtin_clz`) to leverage hardware acceleration. The necessary compiler flags are added to AVX2 and AVX512 variants to enable LZCNT support on x86_64.

* Use 64-bit shifts for offset mask and epoch calculations

Prevents potential undefined behavior or integer overflow by performing bitwise shifts in 64-bit precision before casting the result back to 32 bits. This ensures that the offset mask and maximum epoch are correctly calculated even when the number of offset bits is 32.

+++ src/lib/zxc_common.c
@@ -81,8 +81,8 @@ int zxc_cctx_init(zxc_cctx_t* RESTRICT ctx, const size_t chunk_size, const int m
     ctx->chunk_size = chunk_size;
     const uint32_t offset_bits = zxc_log2_u32((uint32_t)chunk_size);
     ctx->offset_bits = offset_bits;
-    ctx->offset_mask = (1U << offset_bits) - 1;
-    ctx->max_epoch = 1U << (32 - offset_bits);
+    ctx->offset_mask = (uint32_t)((1ULL << offset_bits) - 1);
+    ctx->max_epoch = (uint32_t)(1ULL << (32 - offset_bits));
 
     if (mode == 0) return ZXC_OK;
